Cleanup driver commands after scenario
If a step schedules a lot of driver commands, and an early one
fails, if the driver is retained (i.e. keep the same driver for
all scenarios) it will carry on with those scheduled commands
into the next scenario.

By flushing the command queue after each scenario, we ensure that
no errant commands are run in the wrong steps.